move helper class FakeFacotry to _graph_test_helpers.py
